+++ Features/ArtworkFeature/ArtworkFeatureInterface/ArtworkInterface/Interface.swift
@@ -17,6 +17,4 @@ public protocol ArtworkViewBuildable {
 
 public protocol ArtworkRouting: AnyObject {
     var viewController: UIViewController { get }
-    
-    func pushToUpload()
 }
