+++ Features/PostingFeature/PostingFeature/Source/Upload/UploadViewController.swift
@@ -39,45 +39,3 @@ final class UploadViewController: UIViewController {
     }
 }
 
-public protocol ZiineWebViewPresentable: AnyObject {
-    func getView() -> UIView
-    func loadWebView(urlString: String)
-}
-
-public final class ZiineWebView: UIView, ZiineWebViewPresentable {
-    
-    public func getView() -> UIView {
-        webView
-    }
-    
-    public func loadWebView(urlString: String) {
-        if let url = URL(string: urlString) {
-            let urlRequest = URLRequest(url: url)
-            
-            webView.load(urlRequest)
-        }
-    }
-    
-    // MARK: - Initialize
-    
-    public init() {
-        super.init(frame: .zero)
-        
-        configureUI()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: - UIComponents
-    
-    private let webView = WKWebView()
-    
-    private func configureUI() {
-        addSubview(webView)
-        webView.snp.makeConstraints {
-            $0.edges.equalToSuperview()
-        }
-    }
-}
